requirements.txt
python-dotenv
langchain-together
langchain-community
openai
streamlit
requests
beautifulsoup4
